backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.14.2
google-genai==1.46.0
rembg==2.0.68
onnxruntime==1.23.2
